Version 1.0 Woohoo!

+++ database/channel_operations.py
@@ -1,4 +1,5 @@
 import sqlite3
+import discord
 
 def remove_channel(channel):
     dbname = "databases/"+str(channel.guild.id)+".db"
@@ -6,7 +7,9 @@ def remove_channel(channel):
     db.execute("PRAGMA foreign_keys=ON")
     cursor = db.cursor()
 
-    cursor.execute("DELETE FROM guild_channels WHERE channel_id='"+str(channel.id)+"'")
+    #Removes text channel from database
+    if channel.type == discord.ChannelType.text:
+        cursor.execute("DELETE FROM guild_channels WHERE channel_id='"+str(channel.id)+"'")
 
     db.commit()
     db.close()
@@ -17,7 +20,9 @@ def add_channel(channel):
     db.execute("PRAGMA foreign_keys=ON")
     cursor = db.cursor()
 
-    cursor.execute("INSERT INTO guild_channels (channel_id) VALUES ('"+str(channel.id)+"')")
+    #Adds text channel to database
+    if channel.type == discord.ChannelType.text:
+        cursor.execute("INSERT INTO guild_channels (channel_id) VALUES ('"+str(channel.id)+"')")
 
     db.commit()
     db.close()
